Move rpc_dplx_rec into rpc_dplx_internal
Where it is expected (based upon the prefix).

Clean up includes.

Add _init and _destroy.

Used by future patches.
